+++ src/Cache/Psr16.php
@@ -54,8 +54,16 @@ use Psr\SimpleCache\CacheInterface;
  * @subpackage Caching
  * @internal
  */
+
 final class Psr16 implements DataCache
 {
+    /**
+     * stored PSR-16 cache implementation
+     *
+     * @var CacheInterface
+     */
+    private static $psr16storage = null;
+
     /**
      * PSR-16 cache implementation
      *